founder-copilot/requirements.txt
openai==1.51.2
httpx==0.27.2
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
jinja2==3.1.4
